Bugfix for PointGroupAnalyzer get_symmetry_operations (#889)
Commit f9cc518f14466de92bd58629e92cc86bca5ba193 added a
get_symmetry_operations method to the PointGroupAnalyzer class.
It also introduced a bug, where the symmetry operations returned
only included those found from the initial analysis of a molecule,
instead of the complete set of symmetry operations for the
corresponding point group. This commit fixes this bug, and adds
a more detailed test that checks a full set of symmetry operations
is returned.